requirements.txt
Django==2.1.11
django-filter==2.2.0
djangorestframework==3.10.3
factory-boy==2.12.0
Faker==2.0.4
Markdown==3.1.1
psycopg2==2.8.4
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
text-unidecode==1.3
